Refactor code, extract common and sync modules
